refact(Error): rename exceptions folder to errors
