+++ src/domains/shared/errors/BusinessError.ts
