+++ src/domains/shared/errors/InvalidPassword.ts
