+++ tests/domains/Panel/useCases/Response/CreatePanelResponse.test.ts
@@ -1,5 +1,5 @@
 import CreatePanelResponse from "@/domains/Panel/UseCases/Response/CreatePanelResponse";
-import Error from "@/domains/shared/exceptions/BusinessError";
+import Error from "@/domains/shared/errors/BusinessError";
 
 describe("CreatePanelResponse", () => {
   let response: CreatePanelResponse;
