+++ tests/domains/shared/exceptions/InvalidPassword.test.ts
@@ -1,4 +1,4 @@
-import InvalidPassword from "@/domains/shared/exceptions/InvalidPassword";
+import InvalidPassword from "@/domains/shared/errors/InvalidPassword";
 
 describe("InvalidPassword", () => {
   it("should create an instance of Error", () => {
